finished CPU a b

+++ CPU/run.sh
@@ -1,13 +1,17 @@
 #!/bin/sh
 # please do chmod +x run.sh
-rm -r results
+if [ -d "results" ]; then
+    rm -r results
+fi
 mkdir results
 
 for data_type in iops flops
 do
     for threads in 1 2 4 8 
     do
+        echo "Number of Thread: "$threads", data type: "$data_type
         python cpu.py $threads $data_type >> thread"_"$threads"_type_"$data_type.txt
+        echo "Write to file: thread"_"$threads"_type_"$data_type".txt
 		mv thread"_"$threads"_type_"$data_type.txt results
     done 
 done
